+++ eth.cpp
@@ -70,7 +70,8 @@ ipmi_ret_t GetEthDevice(const uint8_t* reqBuf, uint8_t* replyBuf,
 {
     if ((*dataLen) < sizeof(struct EthDeviceRequest))
     {
-        fprintf(stderr, "Invalid command length: %lu\n", (*dataLen));
+        fprintf(stderr, "Invalid command length: %u\n",
+                static_cast<uint32_t>(*dataLen));
         return IPMI_CC_INVALID;
     }
